backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
openai>=1.3.0
python-dotenv>=0.21.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.33

